refactor: 공통 컴포넌트에 ShoppingCart prefix 추가
